Start creating game board instance
There needs to be an object to control the instance of the tic-tac-toe
game that contains all the relevant game board and player information.
This has been implemented, and since there only needs to be a single
instance of the game board, the module pattern has been used for the
game board. So far, the module instance includes a "blank" 2D array to
represent the board, a Player factory function that is used to create
two players, each with a name and marker, a reference to the current
player, and two functions to retrieve the current player and toggle the
current player, respectively. The module instance only returns these
two functions to aid in gameplay in the future. None of the variables
are returned to prevent them from being modified.