Create function to mark board
When the user selects a square to place their mark in, this needs to
be updated in the array that represents the board. The markBoard
function has been added to do this. When the selected square is input
into the function (in the form of row, column indices), that square
is updated with the current player's mark and the current player is
toggled to the next player. However, this only happens if the given
square is not already occupied. The markBoard function is now returned
with the gameBoard instance.