Check for game over status after each move
If a player makes a winning move or if the game board is entirely
filled, then the game needs to end. There are two new functions
that accomplish this. The first, checkPlayerWin, checks all possible
rows, columns, and diagonals for three in a row after a player move
to check if the player won. The second, checkTie, checks if the board
is filled and there are no more possible moves. Both of these functions
are called after every move, and checkPlayerWin is called first to
make sure the final move of the game wasn't a winning move.