Create function to update player move on display
When the player makes a move, it needs to be update on the game board.
This has been implemented with the help of a display manager object,
implemented with the module pattern since there only needs to be one
display manager. The function takes the row, column, and marker as
inputs and changes the text content of the cell to the correct player
marker. Currently, it can only be called through the console and not
by clicking on the cell on the board.