Create function to update message on display
While the game is still not over, the messsage needs to display whose
turn it is. Once the game is over, the message should display who won
or if the game was a tie. The updateMessage method of the display
manager object now does this. This method is called initially when the
DOM content has loaded and after every time the board is marked.