Add functionality to Reset Game button
When the user clicks the Reset Game button, the board needs to be
cleared and the message needs to be reset. All of the internal
variables of the gameBoard module also need to be reset. This has
been done by adding a call to the resetGame method in the click
event listener of the reset button. Calls to the display manager
have been added to this resetGame method to properly update the
display after the reset.