Add functionality to allow users to input names
There are two player in the game, and each user should be able to
input their name so that the message tells whose turn it is. This has
been implemented with a dialog. The dialog appears on page load and
has two required fields for two user names. When two names are
submitted, the player names are updated on the screen and all of the
messages will contain those user names instead of generic "Player 1"
and "Player 2" names.