Color winning cells red upon game win
As a visual aid to the players, once a player wins a round of the
game, the winning combination is highlighted in red on the display.
This is done with a new method colorCells of the display manager.
This method is also used to reset the cells to black once the game
is reset.